LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-position-inline-002-expected.txt

spacer

012345 789000

PASS .target 1
FAIL .target 2 assert_equals:
<div class="target" style="right: anchor(--a1 right)" data-offset-x="60"></div>
offsetLeft expected 60 but got 100
FAIL .target 3 assert_equals:
<div class="target" style="top: anchor(--a1 top)" data-offset-y="20"></div>
offsetTop expected 20 but got 0
PASS .target 4
